chore(assets): move component assets to src/assets and update imports
